DSP-10 move js to lib folder structure
