+++ lib/colors.js
