+++ pages/search/index.js
@@ -3,9 +3,9 @@ import {
   allFips,
   regionCSUs,
   sortedCSUs,
-} from "../../src/csu.js";
-import dspsXML from "../../src/getXML.js";
-import { removeDuplicates } from "../../src/utils.js";
+} from "../../lib/csu.js";
+import dspsXML from "../../lib/getXML.js";
+import { removeDuplicates } from "../../lib/utils.js";
 const queryString = window.location.search;
 const urlParams = new URLSearchParams(queryString);
 const [serviceName, locationID, languageName] = urlParams.values();
